+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## 0.0.1
+### 0.0.1
 
 * A Flutter package that leverages `json_dynamic_widget` to dynamically render pages from JSON data. This package fetches all the pages from a server when the app launches and provides a function to get a specific page by its name. It also caches the fetched data using `SharedPreferences`.
+
+### 0.0.2
+
+* Included required API response format.
+
+### 0.1.0
+
+* Added `RenderPage` widget to dynamically render JSON data with support for loading indicators and error handling.
+
+### 0.1.2
+
+* Bug fix - resolved conflicts by multiple registry creation.